convenios update working

+++ src/main/java/PROYECTO_TSI/PROYECTO_TSI/CONTROLLERS/ConvenioController.java
@@ -26,6 +26,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.List;
+import java.util.Optional;
 
 @CrossOrigin(origins = "http://localhost:4200",maxAge = 3600)
 @RestController
@@ -43,25 +44,48 @@ public class ConvenioController {
     public ResponseEntity<List<String>> getConvenios() {
         List<String> photos=new ArrayList<String>();
         String filesPath =context.getRealPath("/convenios");
+        //////////////////////////////////////////////
+        List<Convenio>list=convenioService.listar();
+        List<Convenio> lista = new ArrayList<Convenio>();
+        int aux=1;
+        int lenght=list.size();
+        System.out.println("the size is"+lenght);
+        for( Convenio o:list){
+            lista.add(list.get(lenght-aux));
+            aux++;
+        }
+
+        ////////////////////////////////////////
         File filefolder =new File(filesPath);
+        for (final File file  :filefolder.listFiles()){
+            System.out.println("esto es la lista"+ file.getName());
+
+        }
         if (filefolder!=null){
-            for (final File file:filefolder.listFiles()){
-                if(!file.isDirectory()){
-                    String encodeBase64=null;
-                    try{
-                        String extension=FilenameUtils.getExtension(file.getName());
-                        FileInputStream fileInputStream=new FileInputStream(file);
-                        byte[]bytes=new byte[(int)file.length()];
-                        fileInputStream.read(bytes);
-                        encodeBase64= Base64.getEncoder().encodeToString(bytes);
-                        photos.add("data:image/"+extension+";base64,"+encodeBase64);
-                        fileInputStream.close();;
+            for (Convenio o:lista){
+                for (final File file:filefolder.listFiles()){
+                    if (o.getImagen().equals(file.getName())){
+                        if(!file.isDirectory()){
+                            String encodeBase64=null;
+                            try{
+                                String extension=FilenameUtils.getExtension(file.getName());
+                                FileInputStream fileInputStream=new FileInputStream(file);
+                                byte[]bytes=new byte[(int)file.length()];
+                                fileInputStream.read(bytes);
+                                encodeBase64= Base64.getEncoder().encodeToString(bytes);
+                                photos.add("data:image/"+extension+";base64,"+encodeBase64);
+                                fileInputStream.close();;
 
-                    }catch(Exception e){
+                            }catch(Exception e){
 
 
+                            }
+                        }
                     }
-                }
+
+
+            }
+
             }
         }
 
@@ -69,7 +93,17 @@ public class ConvenioController {
     }
     @GetMapping
     public List<Convenio> listar() {
-        return convenioService.listar();
+        List<Convenio>list=convenioService.listar();
+        List<Convenio> lista = new ArrayList<Convenio>();
+        int aux=1;
+        int lenght=list.size();
+        System.out.println("the size is"+lenght);
+        for( Convenio o:list){
+            lista.add(list.get(lenght-aux));
+            aux++;
+        }
+        return lista;
+        //return convenioService.listar();
     }
 
     @PostMapping
@@ -114,53 +148,61 @@ public class ConvenioController {
 
 
     @PostMapping(value = "UpdateConvenioFile")
-    public ResponseEntity<Response> UpdateConvenioFile(@RequestParam("file") MultipartFile file,@RequestParam("convenio")String convenio)throws JsonParseException, JsonMappingException, IOException
+    public ResponseEntity<Response> UpdateConvenioFile(@RequestParam("file") Optional<MultipartFile> file2, @RequestParam("convenio")String convenio)throws JsonParseException, JsonMappingException, IOException
     {
         System.out.println("llega a la funcion");
-
         Convenio convenio1=new ObjectMapper().readValue(convenio,Convenio.class);
-        System.out.println("llega a recibir los datos");
-        System.out.println("nombre anteior"+convenio1.getImagen());
-        System.out.println("nombre nuevo"+file.getOriginalFilename());
+        //System.out.println("llega a recibir los datos");
+        //System.out.println("nombre anteior"+convenio1.getImagen());
+       // System.out.println("nombre nuevo"+file.getOriginalFilename());
+        if (file2.isPresent()){
+            MultipartFile file=file2.get();
+            if (convenio1.getImagen().equals(file.getOriginalFilename()))
+            {
+                System.out.println("entra a imprimir el mismo");
+                convenioService.edit(convenio1);
+                return new ResponseEntity<Response>(new Response("Convenio saved succesfull"), HttpStatus.OK);
+            }
+            else{
+                System.out.println("entra a eliminar");
 
-        if (convenio1.getImagen().equals(file.getOriginalFilename()))
+                String auxiliar = convenio1.getImagen();
+                File fileToDelete = new File("src/main/webApp/convenios/"+auxiliar);
+                System.out.println("this is the name"+fileToDelete.getName());
+                fileToDelete.delete();
+
+                convenio1.setImagen(file.getOriginalFilename());
+
+                boolean isExist = new java.io.File(context.getRealPath("/convenios/")).exists();
+                if(!isExist){
+                    System.out.println("creating directory");
+                    new java.io.File(context.getRealPath("/convenios/")).mkdir();
+                }
+                String filename = file.getOriginalFilename();
+                String modifiedFilename= FilenameUtils.getBaseName(filename)+"_"+System.currentTimeMillis()+"."+FilenameUtils.getExtension(filename);
+                //String modifiedFilename= FilenameUtils.getBaseName(filename)+"."+FilenameUtils.getExtension(filename);
+
+                File serverfile=new java.io.File(context.getRealPath("/convenios/"+ java.io.File.separator+modifiedFilename));
+
+                try{
+                    FileUtils.writeByteArrayToFile(serverfile,file.getBytes());
+                }catch (Exception e){
+                    e.printStackTrace();
+                }
+                convenio1.setImagen(modifiedFilename);
+
+
+                Convenio convenio2= convenioService.edit(convenio1);
+
+            }
+        }
+        else
         {
-            System.out.println("entra a imprimir el mismo");
+            System.out.println("no cambiaron nada");
             convenioService.edit(convenio1);
             return new ResponseEntity<Response>(new Response("Convenio saved succesfull"), HttpStatus.OK);
         }
-        else{
-            System.out.println("entra a eliminar");
 
-            String auxiliar = convenio1.getImagen();
-            File fileToDelete = new File("src/main/webApp/convenios/"+auxiliar);
-            System.out.println("this is the name"+fileToDelete.getName());
-            fileToDelete.delete();
-
-            convenio1.setImagen(file.getOriginalFilename());
-
-            boolean isExist = new java.io.File(context.getRealPath("/convenios/")).exists();
-            if(!isExist){
-                System.out.println("creating directory");
-                new java.io.File(context.getRealPath("/convenios/")).mkdir();
-            }
-            String filename = file.getOriginalFilename();
-            String modifiedFilename= FilenameUtils.getBaseName(filename)+"_"+System.currentTimeMillis()+"."+FilenameUtils.getExtension(filename);
-            //String modifiedFilename= FilenameUtils.getBaseName(filename)+"."+FilenameUtils.getExtension(filename);
-
-            File serverfile=new java.io.File(context.getRealPath("/convenios/"+ java.io.File.separator+modifiedFilename));
-
-            try{
-                FileUtils.writeByteArrayToFile(serverfile,file.getBytes());
-            }catch (Exception e){
-                e.printStackTrace();
-            }
-            convenio1.setImagen(modifiedFilename);
-
-
-            Convenio convenio2= convenioService.edit(convenio1);
-
-        }
 
         if(convenio1!=null){
             return new ResponseEntity<Response>(new Response("convenios saved succesfull"), HttpStatus.OK);
